refactor(layout): RightPanel 컴포넌트로 ChatPanel과 AccountSettingsPanel 분리
- RightPanel: ChatPanel과 AccountSettingsPanel을 모드에 따라 관리하는 wrapper 컴포넌트
- ChatPanel: 순수 채팅 기능만 담당 (계좌 설정 코드 제거)
- AccountSettingsPanel: 독립적인 계좌 설정 UI 컴포넌트
- useRightPanelStore: ChatPanel 상태 관리 (mode: 'chat' | 'account-settings')
- AppShell: RightPanel 사용으로 더 깔끔한 구조
- AppHeader: useRightPanelStore로 계좌 설정 모드 전환
