+++ src/store/useRightPanelStore.js
@@ -1,10 +1,10 @@
 import { create } from 'zustand'
 
-const useChatPanelStore = create((set) => ({
+const useRightPanelStore = create((set) => ({
   mode: 'chat', // 'chat' | 'account-settings'
 
   setChatMode: () => set({ mode: 'chat' }),
   setAccountSettingsMode: () => set({ mode: 'account-settings' }),
 }))
 
-export default useChatPanelStore
+export default useRightPanelStore
